+++ src/Renderer/Latte/Templates/MultipleTemplate.php
@@ -17,17 +17,23 @@ final class MultipleTemplate
     /** @var array<string, scalar|null> */
     public array $elementAttributes;
 
+    /** @var array<string, scalar> */
+    public array $options;
+
     /**
      * @param array<int, Banner>         $banners
      * @param array<string, scalar|null> $elementAttributes
+     * @param array<string, scalar>      $options
      */
     public function __construct(
         Position $position,
         array $banners,
-        array $elementAttributes
+        array $elementAttributes,
+        array $options
     ) {
         $this->position = $position;
         $this->banners = $banners;
         $this->elementAttributes = $elementAttributes;
+        $this->options = $options;
     }
 }