+++ src/Renderer/Latte/Templates/NotFoundTemplate.php
@@ -13,14 +13,20 @@ final class NotFoundTemplate
     /** @var array<string, scalar|null> */
     public array $elementAttributes;
 
+    /** @var array<string, scalar> */
+    public array $options;
+
     /**
      * @param array<string, scalar|null> $elementAttributes
+     * @param array<string, scalar>      $options
      */
     public function __construct(
         Position $position,
-        array $elementAttributes
+        array $elementAttributes,
+        array $options
     ) {
         $this->position = $position;
         $this->elementAttributes = $elementAttributes;
+        $this->options = $options;
     }
 }