+++ tests/Renderer/Latte/LatteRendererBridgeTest.php
@@ -40,11 +40,12 @@ final class LatteRendererBridgeTest extends TestCase
     public function testNotFoundTemplateRendering(
         ResponsePosition $position,
         array $elementAttributes,
+        array $options,
         string $expectationFile
     ): void {
         $renderer = $this->createRendererBridge();
 
-        AssertHtml::assert($expectationFile, $renderer->renderNotFound($position, $elementAttributes));
+        AssertHtml::assert($expectationFile, $renderer->renderNotFound($position, $elementAttributes, $options));
     }
 
     /**
@@ -54,11 +55,12 @@ final class LatteRendererBridgeTest extends TestCase
         ResponsePosition $position,
         ?Banner $banner,
         array $elementAttributes,
+        array $options,
         string $expectationFile
     ): void {
         $renderer = $this->createRendererBridge();
 
-        AssertHtml::assert($expectationFile, $renderer->renderSingle($position, $banner, $elementAttributes));
+        AssertHtml::assert($expectationFile, $renderer->renderSingle($position, $banner, $elementAttributes, $options));
     }
 
     /**
@@ -68,11 +70,12 @@ final class LatteRendererBridgeTest extends TestCase
         ResponsePosition $position,
         ?Banner $banner,
         array $elementAttributes,
+        array $options,
         string $expectationFile
     ): void {
         $renderer = $this->createRendererBridge();
 
-        AssertHtml::assert($expectationFile, $renderer->renderRandom($position, $banner, $elementAttributes));
+        AssertHtml::assert($expectationFile, $renderer->renderRandom($position, $banner, $elementAttributes, $options));
     }
 
     /**
@@ -82,11 +85,12 @@ final class LatteRendererBridgeTest extends TestCase
         ResponsePosition $position,
         array $banners,
         array $elementAttributes,
+        array $options,
         string $expectationFile
     ): void {
         $renderer = $this->createRendererBridge();
 
-        AssertHtml::assert($expectationFile, $renderer->renderMultiple($position, $banners, $elementAttributes));
+        AssertHtml::assert($expectationFile, $renderer->renderMultiple($position, $banners, $elementAttributes, $options));
     }
 
     /**
@@ -95,11 +99,12 @@ final class LatteRendererBridgeTest extends TestCase
     public function testClientSideTemplateRendering(
         RequestPosition $position,
         array $elementAttributes,
+        array $options,
         string $expectationFile
     ): void {
         $renderer = $this->createRendererBridge();
 
-        AssertHtml::assert($expectationFile, $renderer->renderClientSide($position, $elementAttributes));
+        AssertHtml::assert($expectationFile, $renderer->renderClientSide($position, $elementAttributes, $options));
     }
 
     public function notFoundTemplateDataProvider(): array