+++ tests/Renderer/RendererTest.php
@@ -51,7 +51,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderNotFound')
             ->once()
-            ->with($position, [])
+            ->with($position, [], [])
             ->andReturn('not found');
 
         Assert::same('not found', $renderer->render($position));
@@ -75,7 +75,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderSingle')
             ->once()
-            ->with($position, $banner, [])
+            ->with($position, $banner, [], [])
             ->andReturn('single');
 
         Assert::same('single', $renderer->render($position));
@@ -99,7 +99,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderRandom')
             ->once()
-            ->with($position, $banner, [])
+            ->with($position, $banner, [], [])
             ->andReturn('random');
 
         Assert::same('random', $renderer->render($position));
@@ -126,7 +126,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderMultiple')
             ->once()
-            ->with($position, $banners, [])
+            ->with($position, $banners, [], [])
             ->andReturn('multiple');
 
         Assert::same('multiple', $renderer->render($position));
@@ -145,7 +145,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderClientSide')
             ->once()
-            ->with($position, [])
+            ->with($position, [], [])
             ->andReturn('client-side');
 
         Assert::same('client-side', $renderer->renderClientSide($position));
@@ -162,7 +162,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderNotFound')
             ->once()
-            ->with($position, [])
+            ->with($position, [], [])
             ->andThrow(new RendererException('Test exception'));
 
         Assert::exception(
@@ -183,7 +183,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderClientSide')
             ->once()
-            ->with($position, [])
+            ->with($position, [], [])
             ->andThrow(new RendererException('Test exception'));
 
         Assert::exception(
@@ -204,7 +204,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderNotFound')
             ->once()
-            ->with($position, [])
+            ->with($position, [], [])
             ->andThrow(new Exception('Test exception'));
 
         Assert::exception(
@@ -225,7 +225,7 @@ final class RendererTest extends TestCase
         $rendererBridge
             ->shouldReceive('renderClientSide')
             ->once()
-            ->with($position, [])
+            ->with($position, [], [])
             ->andThrow(new Exception('Test exception'));
 
         Assert::exception(